USER/YJS/Backend/requirements.txt
Flask==1.1.2
Flask_SQLAlchemy==2.4.1
requests==2.22.0
beautifulsoup4==4.9.0
python-decouple==3.3
